GTA V Drive (02/09/2020)
Inverted Raw Axis 3. It should return values from -1 to 0 and 0 to 1.
Also removed any confusion from the PWM ports, so that no more errors regarding Error Code -1029 will occur. Removed a bunch of uncessary spacing, and cleaned up a lot of the classes. Plans for next time are to shorten the logic behind several of the methods, as some our code is a little dense.